pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi_example"
version = "0.0.1"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dynamic = ["readme"]
requires-python = ">=3.12"
dependencies = [
    "concurrent-log-handler>=0.9.28",
    "fastapi>=0.128.0",
    "httpx>=0.27.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.8.0",
    "ruff>=0.15.0",
    "typing>=3.10.0.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.1.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=3.0.0",
    "pytest-sugar>=0.9.5",
    "pytest-xdist>=2.5.0",
]

[[project.authors]]
name = "Jonathan van Leeuwen"

[project.urls]
Homepage = "https://github.com/jonathanvanleeuwen/fastapi_example"
"Bug Tracker" = "https://github.com/jonathanvanleeuwen/fastapi_example/issues"

[tool.setuptools]
include-package-data = true
zip-safe = false
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
readme = { file = ["README.md", "LICENSE"], content-type = "text/markdown" }

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = """
--cov src
--cov-report html:reports/htmlcov"""

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint]
disable = [
    "C0114", # missing module docstring
    "C0116", # missing function docstring
    "C0115", # missing class docstring
    "C0301", # line too long
    "W1203",
]

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "missing-function-docstring",
    "missing-class-docstring",
    "missing-module-docstring",
    "line-too-long",
]

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = ["E", "W", "F", "I", "C", "B", "UP"]
lint.ignore = [
    "E501",
    "UP031",
    "W605",
    "G004",
    "D100",
    "D101",
    "D102",
    "D107",
    "C901",  # function too complex
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ['src/fastapi_example/__init__.py:__version__']
branch = "main"
upload_to_repository = false
